requirements.txt
Flask==3.0.0
fastapi==0.116.1
uvicorn==0.35.0
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.0
transformers==4.35.2
Pillow==10.1.0
google-generativeai==0.3.0
nltk==3.8.1
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.33.0
